refactor : test email folder
